Moved assistant entry point to a separate file
