+++ Sources/Extras/CoreMediaIO/DeviceAbstractionLayer/Devices/Sample/Assistant/Server/CMIO_DPA_Sample_Server_Assistant.cpp
@@ -55,9 +55,6 @@
 // Internal Includes
 #include "CMIO_DPA_Sample_Server_Device.h"
 
-// MIG Server Interface
-#include "CMIODPASampleServer.h"
-
 // Public Utility Includes
 #include "CMIODebugMacros.h"
 
@@ -68,40 +65,13 @@
 
 // System Includes
 #include <CoreAudio/HostTime.h>
-#include <mach/mach.h>
-#include <mach/notify.h>
 #include <mach-o/dyld.h>
-#include <servers/bootstrap.h>
-#include <sys/param.h>
 
 // Standard Library Includes
 #include <algorithm>
 
 namespace
 {
-	// As a convenience, use the CMIO::DPA::Sample namespace.  This will allow convienient access in the anonymous namespace as well as in the CMIODPASampleXXX() public interface to the server
-	using namespace CMIO;
-	using namespace CMIO::DPA::Sample;
-
-	//-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-	// MessagesAndNotifications()
-	//	This handles messages from the client and the MACH_NOTIFY_NO_SENDERS notification on the client ports.
-	//-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-	boolean_t MessagesAndNotifications(mach_msg_header_t* request, mach_msg_header_t* reply)
-	{
-		// Invoke the MIG created CMIODPASampleServer() to see if this is one of the client messages it handles
-		boolean_t processed = CMIODPASampleServer(request, reply);
-		
-		// If CMIODPASampleServer() did not process the message see if it is a MACH_NOTIFY_NO_SENDERS notification
-		if (not processed and MACH_NOTIFY_NO_SENDERS == request->msgh_id)
-		{
-			Server::Assistant::Instance()->ClientDied(request->msgh_local_port);
-			processed = true;
-		}
-		
-		return processed;
-	}
-
 	//-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 	// CopyPlugInBundle()
 	//	This locates the com.apple.cmio.DAL.Sample bundle so localized device names can be extracted from it.
@@ -916,60 +886,3 @@ namespace CMIO { namespace DPA { namespace Sample { namespace Server
 		DeviceArrived(*this, iterator);
 	}
 }}}}
-
-// As a convenience, use the CMIO::DPA::Sample::Server namespace
-using namespace CMIO::DPA::Sample::Server;
-
-#pragma mark -
-//---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-// main()
-//---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-int main()
-{
-	// Don't allow any exceptions to escape
-	try
-	{
-		// Check in with the bootstrap port under the agreed upon name to get the servicePort with receive rights
-		mach_port_t servicePort;
-		name_t serviceName = "com.apple.cmio.DPA.Sample";
-		kern_return_t err = bootstrap_check_in(bootstrap_port, serviceName, &servicePort); 
-		if (BOOTSTRAP_SUCCESS != err)
-		{
-			DebugMessage("bootstrap_check_in() failed: 0x%x", err);
-			exit(43);
-		}
-	
-		#if 0
-			// Wait forever until the Debugger can attach to the Assistant process
-			bool waiting = true;
-			while (waiting)
-			{
-				sleep(1);
-			}
-		#endif
-
-		// Add the service port to the Assistant's port set
-		mach_port_t portSet = Assistant::Instance()->GetPortSet();
-		err = mach_port_move_member(mach_task_self(), servicePort, portSet);
-		if (KERN_SUCCESS != err)
-		{
-			DebugMessage("Unable to add service port to port set: 0x%x", err);
-			exit(2);
-		}
-
-		// Service incoming messages from the clients and notifications which were signed up for 
-		while (true)
-		{
-			(void) mach_msg_server(MessagesAndNotifications, 8192, portSet, MACH_MSG_OPTION_NONE);
-		}
-	}
-	catch (const CAException& exception)
-	{
-		exit(exception.GetError());
-	}
-	catch (...)
-	{
-		DebugMessage("Terminated by an an unknown exception");
-		exit(44);
-	}
-}
